Extracted queue objects to their own files
