+++ contrib/mesos/pkg/queue/delay.go
@@ -20,47 +20,7 @@ import (
 	"container/heap"
 	"sync"
 	"time"
-
-	"k8s.io/kubernetes/pkg/util/sets"
 )
-
-type qitem struct {
-	value    interface{}
-	priority Priority
-	index    int
-	readd    func(item *qitem) // re-add the value of the item to the queue
-}
-
-// A priorityQueue implements heap.Interface and holds qitems.
-type priorityQueue []*qitem
-
-func (pq priorityQueue) Len() int { return len(pq) }
-
-func (pq priorityQueue) Less(i, j int) bool {
-	return pq[i].priority.ts.Before(pq[j].priority.ts)
-}
-
-func (pq priorityQueue) Swap(i, j int) {
-	pq[i], pq[j] = pq[j], pq[i]
-	pq[i].index = i
-	pq[j].index = j
-}
-
-func (pq *priorityQueue) Push(x interface{}) {
-	n := len(*pq)
-	item := x.(*qitem)
-	item.index = n
-	*pq = append(*pq, item)
-}
-
-func (pq *priorityQueue) Pop() interface{} {
-	old := *pq
-	n := len(old)
-	item := old[n-1]
-	item.index = -1 // for safety
-	*pq = old[0 : n-1]
-	return item
-}
 
 // concurrency-safe, deadline-oriented queue that returns items after their
 // delay period has expired.
@@ -181,197 +141,4 @@ func (q *DelayQueue) pop(next func() *qitem, cancel <-chan struct{}) interface{}
 		}
 		return x
 	}
-}
-
-// If multiple adds/updates of a single item happen while an item is in the
-// queue before it has been processed, it will only be processed once, and
-// when it is processed, the most recent version will be processed. Items are
-// popped in order of their priority, currently controlled by a delay or
-// deadline assigned to each item in the queue.
-type DelayFIFO struct {
-	// internal deadline-based priority queue
-	delegate *DelayQueue
-	// We depend on the property that items in the set are in the queue and vice versa.
-	items          map[string]*qitem
-	deadlinePolicy DeadlinePolicy
-}
-
-func (q *DelayFIFO) lock() {
-	q.delegate.lock.Lock()
-}
-
-func (q *DelayFIFO) unlock() {
-	q.delegate.lock.Unlock()
-}
-
-func (q *DelayFIFO) rlock() {
-	q.delegate.lock.RLock()
-}
-
-func (q *DelayFIFO) runlock() {
-	q.delegate.lock.RUnlock()
-}
-
-func (q *DelayFIFO) queue() *priorityQueue {
-	return &q.delegate.queue
-}
-
-func (q *DelayFIFO) cond() *sync.Cond {
-	return &q.delegate.cond
-}
-
-// Add inserts an item, and puts it in the queue. The item is only enqueued
-// if it doesn't already exist in the set.
-func (q *DelayFIFO) Add(d UniqueDelayed, rp ReplacementPolicy) {
-	deadline := extractFromDelayed(d)
-	id := d.GetUID()
-	var adder func(*qitem)
-	adder = func(*qitem) {
-		q.add(id, deadline, d, KeepExisting, adder)
-	}
-	q.add(id, deadline, d, rp, adder)
-}
-
-func (q *DelayFIFO) Offer(d UniqueDeadlined, rp ReplacementPolicy) bool {
-	if deadline, ok := extractFromDeadlined(d); ok {
-		id := d.GetUID()
-		q.add(id, deadline, d, rp, func(qp *qitem) { q.Offer(qp.value.(UniqueDeadlined), KeepExisting) })
-		return true
-	}
-	return false
-}
-
-func (q *DelayFIFO) add(id string, deadline Priority, value interface{}, rp ReplacementPolicy, adder func(*qitem)) {
-	q.lock()
-	defer q.unlock()
-	if item, exists := q.items[id]; !exists {
-		item = &qitem{
-			value:    value,
-			priority: deadline,
-			readd:    adder,
-		}
-		heap.Push(q.queue(), item)
-		q.items[id] = item
-	} else {
-		// this is an update of an existing item
-		item.value = rp.replacementValue(item.value, value)
-		item.priority = q.deadlinePolicy.nextDeadline(item.priority, deadline)
-		heap.Fix(q.queue(), item.index)
-	}
-	q.cond().Broadcast()
-}
-
-// Delete removes an item. It doesn't add it to the queue, because
-// this implementation assumes the consumer only cares about the objects,
-// not their priority order.
-func (f *DelayFIFO) Delete(id string) {
-	f.lock()
-	defer f.unlock()
-	delete(f.items, id)
-}
-
-// List returns a list of all the items.
-func (f *DelayFIFO) List() []UniqueID {
-	f.rlock()
-	defer f.runlock()
-	list := make([]UniqueID, 0, len(f.items))
-	for _, item := range f.items {
-		list = append(list, item.value.(UniqueDelayed))
-	}
-	return list
-}
-
-// ContainedIDs returns a stringset.StringSet containing all IDs of the stored items.
-// This is a snapshot of a moment in time, and one should keep in mind that
-// other go routines can add or remove items after you call this.
-func (c *DelayFIFO) ContainedIDs() sets.String {
-	c.rlock()
-	defer c.runlock()
-	set := sets.String{}
-	for id := range c.items {
-		set.Insert(id)
-	}
-	return set
-}
-
-// Get returns the requested item, or sets exists=false.
-func (f *DelayFIFO) Get(id string) (UniqueID, bool) {
-	f.rlock()
-	defer f.runlock()
-	if item, exists := f.items[id]; exists {
-		return item.value.(UniqueID), true
-	}
-	return nil, false
-}
-
-// Variant of DelayQueue.Pop() for UniqueDelayed items
-func (q *DelayFIFO) Await(timeout time.Duration) UniqueID {
-	cancel := make(chan struct{})
-	ch := make(chan interface{}, 1)
-	go func() { ch <- q.pop(cancel) }()
-	var x interface{}
-	select {
-	case <-time.After(timeout):
-		close(cancel)
-		x = <-ch
-	case x = <-ch:
-		// noop
-	}
-	if x != nil {
-		return x.(UniqueID)
-	}
-	return nil
-}
-
-// Variant of DelayQueue.Pop() for UniqueDelayed items
-func (q *DelayFIFO) Pop() UniqueID {
-	return q.pop(nil).(UniqueID)
-}
-
-// variant of DelayQueue.Pop that implements optional cancellation
-func (q *DelayFIFO) pop(cancel chan struct{}) interface{} {
-	next := func() *qitem {
-		q.lock()
-		defer q.unlock()
-		for {
-			for q.queue().Len() == 0 {
-				signal := make(chan struct{})
-				go func() {
-					defer close(signal)
-					q.cond().Wait()
-				}()
-				select {
-				case <-cancel:
-					// we may not have the lock yet, so
-					// broadcast to abort Wait, then
-					// return after lock re-acquisition
-					q.cond().Broadcast()
-					<-signal
-					return nil
-				case <-signal:
-					// we have the lock, re-check
-					// the queue for data...
-				}
-			}
-			x := heap.Pop(q.queue())
-			item := x.(*qitem)
-			unique := item.value.(UniqueID)
-			uid := unique.GetUID()
-			if _, ok := q.items[uid]; !ok {
-				// item was deleted, keep looking
-				continue
-			}
-			delete(q.items, uid)
-			return item
-		}
-	}
-	return q.delegate.pop(next, cancel)
-}
-
-func NewDelayFIFO() *DelayFIFO {
-	f := &DelayFIFO{
-		delegate: NewDelayQueue(),
-		items:    map[string]*qitem{},
-	}
-	return f
 }
