+++ contrib/mesos/pkg/queue/delay_test.go
@@ -28,74 +28,6 @@ const (
 	tolerance = 100 * time.Millisecond // go time delays aren't perfect, this is our tolerance for errors WRT expected timeouts
 )
 
-func timedPriority(t time.Time) Priority {
-	return Priority{ts: t}
-}
-
-func TestPQ(t *testing.T) {
-	t.Parallel()
-
-	var pq priorityQueue
-	if pq.Len() != 0 {
-		t.Fatalf("pq should be empty")
-	}
-
-	now := timedPriority(time.Now())
-	now2 := timedPriority(now.ts.Add(2 * time.Second))
-	pq.Push(&qitem{priority: now2})
-	if pq.Len() != 1 {
-		t.Fatalf("pq.len should be 1")
-	}
-	x := pq.Pop()
-	if x == nil {
-		t.Fatalf("x is nil")
-	}
-	if pq.Len() != 0 {
-		t.Fatalf("pq should be empty")
-	}
-	item := x.(*qitem)
-	if !item.priority.Equal(now2) {
-		t.Fatalf("item.priority != now2")
-	}
-
-	pq.Push(&qitem{priority: now2})
-	pq.Push(&qitem{priority: now2})
-	pq.Push(&qitem{priority: now2})
-	pq.Push(&qitem{priority: now2})
-	pq.Push(&qitem{priority: now2})
-	pq.Pop()
-	pq.Pop()
-	pq.Pop()
-	pq.Pop()
-	pq.Pop()
-	if pq.Len() != 0 {
-		t.Fatalf("pq should be empty")
-	}
-	now4 := timedPriority(now.ts.Add(4 * time.Second))
-	now6 := timedPriority(now.ts.Add(4 * time.Second))
-	pq.Push(&qitem{priority: now2})
-	pq.Push(&qitem{priority: now4})
-	pq.Push(&qitem{priority: now6})
-	pq.Swap(0, 2)
-	if !pq[0].priority.Equal(now6) || !pq[2].priority.Equal(now2) {
-		t.Fatalf("swap failed")
-	}
-	if pq.Less(1, 2) {
-		t.Fatalf("now4 < now2")
-	}
-}
-
-func TestPopEmptyPQ(t *testing.T) {
-	t.Parallel()
-	defer func() {
-		if r := recover(); r == nil {
-			t.Fatalf("Expected panic from popping an empty PQ")
-		}
-	}()
-	var pq priorityQueue
-	pq.Pop()
-}
-
 type testjob struct {
 	d        time.Duration
 	t        time.Time
@@ -329,80 +261,4 @@ func TestDQ_negative_delay(t *testing.T) {
 	if item.d != delay {
 		t.Fatalf("d != delay")
 	}
-}
-
-func TestDFIFO_sanity_check(t *testing.T) {
-	t.Parallel()
-	assert := assert.New(t)
-
-	df := NewDelayFIFO()
-	delay := 2 * time.Second
-	df.Add(&testjob{d: delay, uid: "a", instance: 1}, ReplaceExisting)
-	assert.True(df.ContainedIDs().Has("a"))
-
-	// re-add by ReplaceExisting
-	df.Add(&testjob{d: delay, uid: "a", instance: 2}, ReplaceExisting)
-	assert.True(df.ContainedIDs().Has("a"))
-
-	a, ok := df.Get("a")
-	assert.True(ok)
-	assert.Equal(a.(*testjob).instance, 2)
-
-	// re-add by KeepExisting
-	df.Add(&testjob{d: delay, uid: "a", instance: 3}, KeepExisting)
-	assert.True(df.ContainedIDs().Has("a"))
-
-	a, ok = df.Get("a")
-	assert.True(ok)
-	assert.Equal(a.(*testjob).instance, 2)
-
-	// pop last
-	before := time.Now()
-	x := df.Pop()
-	assert.Equal(a.(*testjob).instance, 2)
-
-	now := time.Now()
-	waitPeriod := now.Sub(before)
-
-	if waitPeriod+tolerance < delay {
-		t.Fatalf("delay too short: %v, expected: %v", waitPeriod, delay)
-	}
-	if x == nil {
-		t.Fatalf("x is nil")
-	}
-	item := x.(*testjob)
-	if item.d != delay {
-		t.Fatalf("d != delay")
-	}
-}
-
-func TestDFIFO_Offer(t *testing.T) {
-	t.Parallel()
-	assert := assert.New(t)
-
-	dq := NewDelayFIFO()
-	delay := time.Second
-
-	added := dq.Offer(&testjob{instance: 1}, ReplaceExisting)
-	if added {
-		t.Fatalf("DelayFIFO should not add offered job without deadline")
-	}
-
-	deadline := time.Now().Add(delay)
-	added = dq.Offer(&testjob{deadline: &deadline, instance: 2}, ReplaceExisting)
-	if !added {
-		t.Fatalf("DelayFIFO should add offered job with deadline")
-	}
-
-	before := time.Now()
-	x := dq.Pop()
-
-	now := time.Now()
-	waitPeriod := now.Sub(before)
-
-	if waitPeriod+tolerance < delay {
-		t.Fatalf("delay too short: %v, expected: %v", waitPeriod, delay)
-	}
-	assert.NotNil(x)
-	assert.Equal(x.(*testjob).instance, 2)
 }
